requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
requests>=2.31.0
prophet>=1.1.0
dash>=2.14.0
plotly>=5.18.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
openai>=1.0.0
gunicorn>=21.2.0
PyJWT>=2.8.0
bcrypt>=4.1.0
fpdf2>=2.7.0
kaleido>=0.2.1
beautifulsoup4>=4.12.0
